Move windows module tests to pytest
